+++ src/provider/zeebe/ZeebePropertiesProvider.js
@@ -5,6 +5,7 @@ import { findIndex } from 'min-dash';
 import {
   ActiveElementsProps,
   AdHocCompletionProps,
+  AdHocSubProcessImplementationProps,
   AssignmentDefinitionProps,
   BusinessRuleImplementationProps,
   CalledDecisionProps,
@@ -18,6 +19,7 @@ import {
   InputProps,
   MessageProps,
   MultiInstanceProps,
+  OutputCollectionProps,
   OutputPropagationProps,
   OutputProps,
   PriorityDefinitionProps,
@@ -35,7 +37,11 @@ import {
 
 import { ExtensionPropertiesProps } from '../shared/ExtensionPropertiesProps';
 
-import { isMessageEndEvent, isMessageThrowEvent } from './utils/ZeebeServiceTaskUtil';
+import {
+  isMessageEndEvent,
+  isMessageThrowEvent,
+  isZeebeServiceTask
+} from './utils/ZeebeServiceTaskUtil';
 
 /**
  * @typedef { import('@bpmn-io/properties-panel').EntryDefinition } Entry
@@ -48,6 +54,8 @@ const ZEEBE_GROUPS = [
   CalledDecisionGroup,
   ScriptImplementationGroup,
   ScriptGroup,
+  OutputCollectionGroup,
+  AdHocSubProcessImplementationGroup,
   UserTaskImplementationGroup,
   TaskDefinitionGroup,
   AssignmentDefinitionGroup,
@@ -89,8 +97,9 @@ export default class ZeebePropertiesProvider {
       updateMultiInstanceGroup(groups, element);
       updateAdHocCompletionGroup(groups, element);
 
-      // (3) remove message group when not applicable
+      // (3) remove groups when not applicable
       groups = removeMessageGroup(groups, element);
+      groups = removeCompletionGroup(groups, element);
 
       return groups;
     };
@@ -106,6 +115,20 @@ export default class ZeebePropertiesProvider {
 
 ZeebePropertiesProvider.$inject = [ 'propertiesPanel', 'injector' ];
 
+
+function AdHocSubProcessImplementationGroup(element, injector) {
+  const translate = injector.get('translate');
+  const group = {
+    id: 'adHocSubProcessImplementation',
+    label: translate('Implementation'),
+    entries: [
+      ...AdHocSubProcessImplementationProps({ element })
+    ],
+    component: Group
+  };
+
+  return group.entries.length ? group : null;
+}
 
 function CalledDecisionGroup(element, injector) {
   const translate = injector.get('translate');
@@ -313,6 +336,20 @@ function AssignmentDefinitionGroup(element, injector) {
   return group.entries.length ? group : null;
 }
 
+function OutputCollectionGroup(element, injector) {
+  const translate = injector.get('translate');
+  const group = {
+    id: 'outputCollection',
+    label: translate('Output collection'),
+    entries: [
+      ...OutputCollectionProps({ element })
+    ],
+    component: Group
+  };
+
+  return group.entries.length ? group : null;
+}
+
 function ActiveElementsGroup(element, injector) {
   const translate = injector.get('translate');
   const group = {
@@ -481,10 +518,10 @@ function updateAdHocCompletionGroup(groups, element) {
     AdHocCompletionProps({ element })
   );
 
-  // reorder groups to move active elements before adHoc completion
+  // reorder groups to move adHoc completion after active elements
   const activeElementsGroup = findGroup(groups, 'activeElements');
   if (activeElementsGroup) {
-    reorderGroupsIfNecessary(groups, activeElementsGroup, adHocCompletionGroup);
+    moveGroupAfter(groups, adHocCompletionGroup, activeElementsGroup);
   }
 }
 
@@ -499,30 +536,20 @@ function removeMessageGroup(groups, element) {
   return groups;
 }
 
+// remove completion group if AdHoc SubProcess is implemented as job worker
+function removeCompletionGroup(groups, element) {
+  if (isZeebeServiceTask(element)) {
+    return groups.filter(group => group.id !== 'adHocCompletion');
+  }
+
+  return groups;
+}
+
 
 // helper /////////////////////
 
 function findGroup(groups, id) {
   return groups.find(g => g.id === id);
-}
-
-/**
- * Put groups into the defined order if necessary.
- *
- * @param {Group[]} groups
- * @param {Group} firstGroup
- * @param {Group} secondGroup
- */
-function reorderGroupsIfNecessary(groups, firstGroup, secondGroup) {
-  const firstGroupIndex = groups.indexOf(firstGroup);
-  const secondGroupIndex = groups.indexOf(secondGroup);
-
-  if (firstGroupIndex === -1 || secondGroupIndex === -1 || firstGroupIndex <= secondGroupIndex) {
-    return;
-  }
-
-  groups[firstGroupIndex] = secondGroup;
-  groups[secondGroupIndex] = firstGroup;
 }
 
 /**
@@ -584,4 +611,15 @@ function indexEntriesById(entries) {
     index[entry.id] = idx;
     return index;
   }, {});
+}
+
+/**
+ * Move a group after a reference group. Assumes that both groups are present.
+ */
+function moveGroupAfter(groups, groupToMove, referenceGroup) {
+  const groupIndex = groups.indexOf(groupToMove);
+  groups.splice(groupIndex, 1);
+
+  const referenceIndex = groups.indexOf(referenceGroup);
+  groups.splice(referenceIndex + 1, 0, groupToMove);
 }